Safari-609.1.11.1/LayoutTests/fast/dom/getElementsByClassName/return-type-expected.txt

Tests that getElementsByTagName() returns an HTMLCollection.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementsByClassName("test").__proto__ is HTMLCollection.prototype
PASS document.body.getElementsByClassName("test").__proto__ is HTMLCollection.prototype
PASS successfullyParsed is true

TEST COMPLETE
